+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 **NOTE**: As permitted by semantic versioning, backward compatibility is NOT maintained for initial development, i.e. releases before 1.0.0.
 
-## Unreleased
+## [0.5.2] - 2019-02-02
 ### Added
 - The `warn-octal-off` flag.
 - The `info` command now accepts one or more argument for specific aspects of the program state.